processor-api/requirements.txt
# FastAPI e dependências
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0  # Para BaseSettings
email-validator>=2.0.0

# Processamento de dados
pandas>=2.0.0
numpy>=1.24.3
scikit-learn>=1.3.0

# Cliente HTTP para comunicação entre serviços
httpx>=0.24.0

# Banco de dados
sqlalchemy>=2.0.0
asyncpg>=0.27.0  # Driver PostgreSQL assíncrono
alembic>=1.10.0  # Migrações de banco de dados

# Logging e monitoramento
structlog>=23.1.0

cafe-autofe>=0.1.0

pytest>=7.3.1
pytest-asyncio>=0.21.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0